+++ src/Pim/Product/ProductDeclinationUpsertData.php
@@ -43,6 +43,9 @@ final class ProductDeclinationUpsertData
     /** @var array */
     private $priceTiers;
 
+    /** @var null|string */
+    protected $supplierReference;
+
     /**
      * ProductDeclinationUpsertData constructor.
      *
@@ -182,6 +185,7 @@ final class ProductDeclinationUpsertData
             'affiliateLink',
             'infiniteStock',
             'priceTiers',
+            'supplierReference',
         ];
         foreach ($metadata->getReflectionClass()->getProperties() as $prop) {
             if (!\in_array($prop->getName(), $nullableProperties)) {
@@ -214,6 +218,28 @@ final class ProductDeclinationUpsertData
             $data['infinite_stock'] = $this->infiniteStock;
         }
 
+        if ($this->supplierReference !== null) {
+            $data['supplier_reference'] = $this->supplierReference;
+        }
+
         return $data;
+    }
+
+    /** @return string|null */
+    public function getSupplierReference(): ?string
+    {
+        return $this->supplierReference;
+    }
+
+    /**
+     * @param string|null $supplierReference
+     *
+     * @return ProductDeclinationUpsertData
+     */
+    public function setSupplierReference(?string $supplierReference): self
+    {
+        $this->supplierReference = $supplierReference;
+
+        return $this;
     }
 }